// js/frontend/src/lib/types/QrCodeParams.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ImageFormats } from "./ImageFormats";

export interface QrCodeParams {
  format?: ImageFormats;
  bg_red?: number;
  bg_green?: number;
  bg_blue?: number;
  bg_alpha?: number;
  fg_red?: number;
  fg_green?: number;
  fg_blue?: number;
  fg_alpha?: number;
}